fix encoding

+++ asd/labels.py
@@ -25,7 +25,7 @@ class OHELabelTransformer(th.nn.Module):
             A tensor containing the one-hot encoded labels.
         """
         # Convert labels to one-hot encoding
-        one_hot = F.one_hot(y.to(th.int64), num_classes=self.num_classes)
+        one_hot = F.one_hot(y.to(th.int64), num_classes=self.num_classes).to(th.float32)
 
         return one_hot
